scripts: common_utils: add ASK_USER function
Prompts the user for a yes/no response and returns true if the answer is affirmative.

+++ scripts/tools/setup_gitconfig.sh
@@ -3,8 +3,13 @@ CONFIG_FILE="$HOME/.gitconfig"
 LOG_STEP_IN true "Set up gitconfig"
 
 if [[ -f "$CONFIG_FILE" ]]; then
-    LOGW "- Existing .gitconfig found, removing..."
-    rm -f "$CONFIG_FILE"
+    LOGW "- Existing .gitconfig found"
+    if ASK_USER "Overwrite existing config?"; then
+        LOG "- Removing old .gitconfig file..."
+        rm -f "$CONFIG_FILE"
+    else
+        LOG "- Skipping overwrite of existing config..."
+    fi
 fi
 
 LOG "- Applying new git configuration"
